requirements.txt
black
essentia_tensorflow
numpy
tqdm
